WIP: Sanitize using unsigned-integer-overflow and safe-stack.
This enables unsigned-integer-overflow and safe-stack runtime
sanitizers on the shared libraries libssl and libcrypto, for
platform (as opposed to unbundled) builds only. These
sanitizers have very low (1-2%) speed overhead and will protect
against unsigned integer overflows and clumsy stack smashing by
aborting.

+++ src/crypto/asn1/a_int.c
@@ -367,7 +367,7 @@ int ASN1_INTEGER_set(ASN1_INTEGER *a, long v)
 		d>>=8;
 		}
 	j=0;
-	for (k=i-1; k >=0; k--)
+	for (k=(int)i-1; k >=0; k--)
 		a->data[j++]=buf[k];
 	a->length=j;
 	return(1);
